move rinfo command to archiver.rinfo
